signal: make windows docs for signal module show up on unix builds (#3770)
